Fix icons_rc and shortcut icon
