+++ foqus_lib/gui/flowsheet/dataBrowserDialog_UI.ui
@@ -16,7 +16,7 @@
   <layout class="QVBoxLayout" name="verticalLayout_2"/>
  </widget>
  <resources>
-  <include location="../icons.qrc"/>
+  <include location="foqus_lib.gui.icons.qrc"/>
  </resources>
  <connections/>
 </ui>
