+++ foqus_lib/gui/flowsheet/dataFilterDialog_UI.ui
@@ -166,7 +166,7 @@
   </layout>
  </widget>
  <resources>
-  <include location="../icons.qrc"/>
+  <include location="foqus_lib.gui.icons.qrc"/>
  </resources>
  <connections/>
 </ui>
