+++ foqus_lib/gui/flowsheet/flowsheetSettingsDialog_UI.ui
@@ -394,7 +394,7 @@
   </layout>
  </widget>
  <resources>
-  <include location="icons.qrc"/>
+  <include location="foqus_lib.gui.icons.qrc"/>
  </resources>
  <connections/>
 </ui>
