+++ foqus_lib/gui/flowsheet/nodePanel_UI.ui
@@ -1285,7 +1285,7 @@
   </widget>
  </widget>
  <resources>
-  <include location="../icons.qrc"/>
+  <include location="foqus_lib.gui.icons.qrc"/>
  </resources>
  <connections/>
 </ui>
