+++ foqus_lib/gui/help/helpBrowser_UI.ui
@@ -354,7 +354,7 @@
   </widget>
  </widget>
  <resources>
-  <include location="../icons.qrc"/>
+  <include location="foqus_lib.gui.icons.qrc"/>
  </resources>
  <connections/>
 </ui>
