+++ foqus_lib/gui/main/sessionDescriptionEdit_UI.ui
@@ -261,7 +261,7 @@
   </layout>
  </widget>
  <resources>
-  <include location="../icons.qrc"/>
+  <include location="foqus_lib.gui.icons.qrc"/>
  </resources>
  <connections>
   <connection>
