+++ foqus_lib/gui/main/settingsFrame_UI.ui
@@ -1246,7 +1246,7 @@
   </layout>
  </widget>
  <resources>
-  <include location="../icons.qrc"/>
+  <include location="foqus_lib.gui.icons.qrc"/>
  </resources>
  <connections>
   <connection>
