+++ foqus_lib/gui/main/turbineConfig_UI.ui
@@ -196,7 +196,7 @@
   <tabstop>verBox</tabstop>
  </tabstops>
  <resources>
-  <include location="../icons.qrc"/>
+  <include location="foqus_lib.gui.icons.qrc"/>
  </resources>
  <connections/>
 </ui>
