+++ foqus_lib/gui/model/dmfUploadDialog_UI.ui
@@ -338,7 +338,7 @@
   </layout>
  </widget>
  <resources>
-  <include location="../icons.qrc"/>
+  <include location="foqus_lib.gui.icons.qrc"/>
  </resources>
  <connections/>
 </ui>
