+++ foqus_lib/gui/model/gatewayUploadDialog_UI.ui
@@ -417,7 +417,7 @@
   </layout>
  </widget>
  <resources>
-  <include location="../icons.qrc"/>
+  <include location="foqus_lib.gui.icons.qrc"/>
  </resources>
  <connections/>
 </ui>
