+++ foqus_lib/gui/optimization/optMonitor_UI.ui
@@ -225,7 +225,7 @@
   </layout>
  </widget>
  <resources>
-  <include location="../icons.qrc"/>
+  <include location="foqus_lib.gui.icons.qrc"/>
  </resources>
  <connections/>
 </ui>
