+++ foqus_lib/gui/optimization/optSampleGenDialog_UI.ui
@@ -163,7 +163,7 @@
   </layout>
  </widget>
  <resources>
-  <include location="../icons.qrc"/>
+  <include location="foqus_lib.gui.icons.qrc"/>
  </resources>
  <connections>
   <connection>
