+++ foqus_lib/gui/optimization/optSetupFrame_UI.ui
@@ -759,7 +759,7 @@ def objfunc(x, f, fail):
   </layout>
  </widget>
  <resources>
-  <include location="../icons.qrc"/>
+  <include location="foqus_lib.gui.icons.qrc"/>
  </resources>
  <connections>
   <connection>
