+++ foqus_lib/gui/surrogate/surrogateFrame_UI.ui
@@ -337,7 +337,7 @@
   </layout>
  </widget>
  <resources>
-  <include location="../icons.qrc"/>
+  <include location="foqus_lib.gui.icons.qrc"/>
  </resources>
  <connections/>
 </ui>
