--- icons_rc.py
@@ -1,1 +1,0 @@
-import foqus_lib.gui.icons_rc as icons_rc
